More warnings removed from MIND on Visual Studio

+++ Libraries/MINDful/src/simd/emulated/integers/SIMDVector4ui.cpp
@@ -45,10 +45,10 @@ namespace Mind {
 
         Vector4ui::Vector4ui(const float32x4_t& vec) {
             float32x4_t vecLocal = vec ;
-            m_inner[0] = vecLocal[0] ;
-            m_inner[1] = vecLocal[1] ;
-            m_inner[2] = vecLocal[2] ;
-            m_inner[3] = vecLocal[3] ;
+            m_inner[0] = static_cast<uint32_t>(vecLocal[0]) ;
+            m_inner[1] = static_cast<uint32_t>(vecLocal[1]) ;
+            m_inner[2] = static_cast<uint32_t>(vecLocal[2]) ;
+            m_inner[3] = static_cast<uint32_t>(vecLocal[3]) ;
         }
 
         Vector4ui::~Vector4ui() {}
